long -> int

+++ Q3E/src/main/jni/doom3/neo/prey/gamesys/Class.cpp
@@ -260,7 +260,7 @@ idClass::FindUninitializedMemory
 */
 void idClass::FindUninitializedMemory( void ) {
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-	unsigned long *ptr = ( ( unsigned long * )this ) - 1;
+	unsigned int *ptr = ( ( unsigned int * )this ) - 1; // 64long
 	int size = *ptr;
 	assert( ( size & 3 ) == 0 );
 	size >>= 2;
@@ -429,7 +429,7 @@ void * idClass::operator new( size_t s ) {
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-	unsigned long *ptr = (unsigned long *)p;
+	unsigned int *ptr = (unsigned int *)p; // 64long
 	int size = s;
 	assert( ( size & 3 ) == 0 );
 	// HUMANHEAD tmj: bugfix - shifting by 2 gives the number of DWORDs to fill.
@@ -453,7 +453,7 @@ void * idClass::operator new( size_t s, int, int, char *, int ) {
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-	unsigned long *ptr = (unsigned long *)p;
+	unsigned int *ptr = (unsigned int *)p; // 64long
 	int size = s;
 	assert( ( size & 3 ) == 0 );
 	// HUMANHEAD tmj: bugfix - shifting by 2 gives the number of DWORDs to fill.
